payment_service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
payos>=1.0.6